requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
boto3==1.37.23
botocore==1.37.23
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coverage==7.6.12
crispy-bootstrap5==2024.10
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.6.0
django-crispy-forms==2.3
django-simple-history==3.8.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
mccabe==0.7.0
narwhals==1.28.0
numpy==2.2.3
openai==1.69.0
packaging==24.2
pillow==11.1.0
plotly==6.0.0
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.11.1
pydantic_core==2.33.0
pyflakes==3.2.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
RapidFuzz==3.12.1
requests==2.32.3
s3transfer==0.11.4
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
